+++ sinch/domains/conversation/models/v1/messages/types/display_mode_type.py
@@ -1,0 +1,8 @@
+from typing import Literal, Union
+
+from pydantic import StrictStr
+
+DisplayModeType = Union[
+    Literal["DISPLAY_MODE_UNSPECIFIED", "PERSISTENT"],
+    StrictStr,
+]