+++ tests/unit/domains/numbers/v1/sinch_events/test_number_sinch_event.py
@@ -2,6 +2,8 @@ from datetime import datetime, timezone
 import pytest
 from sinch.domains.numbers.sinch_events.v1 import SinchEvents
 from sinch.domains.numbers.sinch_events.v1.events import NumberSinchEvent
+from sinch.domains.numbers.sinch_events.v1.events.number_order_sinch_event import NumberOrderSinchEvent
+from sinch.domains.numbers.sinch_events.v1.events.active_number_sinch_event import ActiveNumberSinchEvent
 
 
 @pytest.fixture
@@ -20,7 +22,7 @@ def sinch_events():
 
 
 @pytest.fixture
-def base_payload_parse_event():
+def base_payload_parse_active_number():
     return {
         "eventId": "01jr7stexp0znky34pj07dwp41",
         "projectId": "project-id",
@@ -32,6 +34,36 @@ def base_payload_parse_event():
         "internalFailureCode": None,
         "timestamp": "2025-04-07T09:38:04.854087603"
     }
+
+@pytest.fixture
+def base_payload_parse_number_order():
+    return {
+        "eventId": "01jr7stexp0znky34pj07dwp41",
+        "projectId": "project-id",
+        "resourceId": "+1234567890",
+        "resourceType": "NUMBER_ORDER",
+        "eventType": "NUMBER_ORDER_PROCESSING", 
+        "status": "SUCCEEDED",
+        "failureCode": None,
+        "internalFailureCode": None,
+        "timestamp": "2025-04-07T09:38:04.854087603"
+    }
+
+@pytest.fixture
+def base_payload_parse_number_sinch_vent():
+    return {
+        "eventId": "01jr7stexp0znky34pj07dwp41",
+        "projectId": "project-id",
+        "resourceId": "+1234567890",
+        "resourceType": "NEW_EVENT",
+        "eventType": "NUMBER_ORDER_PROCESSING", 
+        "status": "SUCCEEDED",
+        "failureCode": None,
+        "internalFailureCode": None,
+        "timestamp": "2025-04-07T09:38:04.854087603"
+    }
+
+
 
 
 def test_valid_signature_header_expects_successful_validation(sinch_events, string_to_sign):
@@ -65,9 +97,10 @@ def test_parse_event_expects_timestamp_as_utc(sinch_events, test_name, timestamp
     assert parsed.timestamp == expected
 
 
-def test_parse_event_expects_parsed_response(sinch_events, base_payload_parse_event):
-    response = sinch_events.parse_event(base_payload_parse_event)
+def test_parse_event_expects_parsed_active_number_response(sinch_events, base_payload_parse_active_number):
+    response = sinch_events.parse_event(base_payload_parse_active_number)
     assert isinstance(response, NumberSinchEvent)
+    assert isinstance(response, ActiveNumberSinchEvent)
     assert response.event_id == "01jr7stexp0znky34pj07dwp41"
     assert response.project_id == "project-id"
     assert response.resource_id == "+1234567890"
@@ -75,6 +108,41 @@ def test_parse_event_expects_parsed_response(sinch_events, base_payload_parse_ev
     assert response.event_type == "PROVISIONING_TO_VOICE_PLATFORM"
     assert response.status == "SUCCEEDED"
     assert response.failure_code is None
+    assert response.internal_failure_code is None
+    expected_timestamp = datetime(
+        2025, 4, 7, 9, 38, 4, 854087, tzinfo=timezone.utc
+    )
+    assert response.timestamp == expected_timestamp
+
+
+def test_parse_event_expects_parsed_number_order_response(sinch_events, base_payload_parse_number_order):
+    response = sinch_events.parse_event(base_payload_parse_number_order)
+    assert isinstance(response, NumberSinchEvent)
+    assert isinstance(response, NumberOrderSinchEvent)
+    assert response.resource_type == "NUMBER_ORDER"
+    assert response.event_id == "01jr7stexp0znky34pj07dwp41"
+    assert response.project_id == "project-id"
+    assert response.resource_id == "+1234567890"
+    assert response.resource_type == "NUMBER_ORDER"
+    assert response.event_type == "NUMBER_ORDER_PROCESSING"
+    assert response.status == "SUCCEEDED"
+    assert response.failure_code is None #Tradeoff because of the inheritance of NumberSinchEvent
+    assert response.internal_failure_code is None #Tradeoff because of the inheritance of NumberSinchEvent
+    expected_timestamp = datetime(
+        2025, 4, 7, 9, 38, 4, 854087, tzinfo=timezone.utc
+    )
+    assert response.timestamp == expected_timestamp
+
+def test_parse_event_expects_parsed_number_sinch_event_response(sinch_events, base_payload_parse_number_sinch_vent):
+    response = sinch_events.parse_event(base_payload_parse_number_sinch_vent)
+    assert isinstance(response, NumberSinchEvent)
+    assert response.event_id == "01jr7stexp0znky34pj07dwp41"
+    assert response.project_id == "project-id"
+    assert response.resource_id == "+1234567890"
+    assert response.resource_type == "NEW_EVENT"
+    assert response.event_type == "NUMBER_ORDER_PROCESSING"
+    assert response.status == "SUCCEEDED"
+    assert response.failure_code is None 
     assert response.internal_failure_code is None
     expected_timestamp = datetime(
         2025, 4, 7, 9, 38, 4, 854087, tzinfo=timezone.utc